pyproject.toml
[project]
name = "multi-agent-travel-planner"
version = "0.1.0"
description = "FastAPI MVP for a mock multi-agent travel planner"
requires-python = ">=3.10,<3.11"
dependencies = [
  "fastapi==0.135.2",
  "uvicorn[standard]==0.41.0",
  "pydantic==2.12.5",
  "pydantic-settings==2.13.1",
  "pyyaml>=6.0.0",
  "sqlalchemy>=2.0.0",
  "pymysql>=1.1.0",
  "redis>=5.0.0",
  "mcp>=1.26.0",
  "qwen-agent[mcp]>=0.0.33",
  "soundfile>=0.13.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx==0.28.1"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
